Fix padding on call emergency services button
Why:
Current the padding on the call emergency services button is not
preserved if the button text is too long.
